Update to Opal 0.8.x plugin syntax and location. Add an actual changelog.
