+++ CHANGELOG.md
@@ -1,3 +1,16 @@
-Changelog: 
-		
-		
+### 0.3 Release
+
+Significant API churn with contextualising and defining Pathways, Steps, et cetera.
+See updated README for further details.
+
+Targets Opal 0.8.x
+
+### 0.2 Release
+
+Targets Opal 0.7.x
+
+### 0.1 Release
+
+Initial release, still very alpha.
+
+Targets Opal 0.6.x
